requirements.txt
paho-mqtt>=1.6.1
matplotlib>=3.5.0
numpy>=1.20.0
PySide6>=6.2.0 